all functions now run through game object
