+++ snowman.js
@@ -1,794 +1,75 @@
 const readLineSync = require("readline-sync");
-const readline = require("readline-sync");
-const dictionary = [
-  "able",
-  "about",
-  "account",
-  "acid",
-  "across",
-  "addition",
-  "adjustment",
-  "advertisement",
-  "after",
-  "again",
-  "against",
-  "agreement",
-  "almost",
-  "among",
-  "amount",
-  "amusement",
-  "angle",
-  "angry",
-  "animal",
-  "answer",
-  "apparatus",
-  "apple",
-  "approval",
-  "arch",
-  "argument",
-  "army",
-  "attack",
-  "attempt",
-  "attention",
-  "attraction",
-  "authority",
-  "automatic",
-  "awake",
-  "baby",
-  "back",
-  "balance",
-  "ball",
-  "band",
-  "base",
-  "basin",
-  "basket",
-  "bath",
-  "beautiful",
-  "because",
-  "before",
-  "behaviour",
-  "belief",
-  "bell",
-  "bent",
-  "berry",
-  "between",
-  "bird",
-  "birth",
-  "bite",
-  "bitter",
-  "black",
-  "blade",
-  "blood",
-  "blow",
-  "blue",
-  "board",
-  "boat",
-  "body",
-  "boiling",
-  "bone",
-  "book",
-  "boot",
-  "bottle",
-  "brain",
-  "brake",
-  "branch",
-  "brass",
-  "bread",
-  "breath",
-  "brick",
-  "bridge",
-  "bright",
-  "broken",
-  "brother",
-  "brown",
-  "brush",
-  "bucket",
-  "building",
-  "bulb",
-  "burn",
-  "burst",
-  "business",
-  "butter",
-  "button",
-  "cake",
-  "camera",
-  "canvas",
-  "card",
-  "care",
-  "carriage",
-  "cart",
-  "cause",
-  "certain",
-  "chain",
-  "chalk",
-  "chance",
-  "change",
-  "cheap",
-  "cheese",
-  "chemical",
-  "chest",
-  "chief",
-  "chin",
-  "church",
-  "circle",
-  "clean",
-  "clear",
-  "clock",
-  "cloth",
-  "cloud",
-  "coal",
-  "coat",
-  "cold",
-  "collar",
-  "colour",
-  "comb",
-  "come",
-  "comfort",
-  "committee",
-  "common",
-  "company",
-  "comparison",
-  "competition",
-  "complete",
-  "complex",
-  "condition",
-  "connection",
-  "conscious",
-  "control",
-  "cook",
-  "copper",
-  "copy",
-  "cord",
-  "cork",
-  "cotton",
-  "cough",
-  "country",
-  "cover",
-  "crack",
-  "credit",
-  "crime",
-  "cruel",
-  "crush",
-  "current",
-  "curtain",
-  "curve",
-  "cushion",
-  "damage",
-  "danger",
-  "dark",
-  "daughter",
-  "dead",
-  "dear",
-  "death",
-  "debt",
-  "decision",
-  "deep",
-  "degree",
-  "delicate",
-  "dependent",
-  "design",
-  "desire",
-  "destruction",
-  "detail",
-  "development",
-  "different",
-  "digestion",
-  "direction",
-  "dirty",
-  "discovery",
-  "discussion",
-  "disease",
-  "disgust",
-  "distance",
-  "distribution",
-  "division",
-  "door",
-  "doubt",
-  "down",
-  "drain",
-  "drawer",
-  "dress",
-  "drink",
-  "driving",
-  "drop",
-  "dust",
-  "early",
-  "earth",
-  "east",
-  "edge",
-  "education",
-  "effect",
-  "elastic",
-  "electric",
-  "engine",
-  "enough",
-  "equal",
-  "error",
-  "even",
-  "event",
-  "ever",
-  "every",
-  "example",
-  "exchange",
-  "existence",
-  "expansion",
-  "experience",
-  "expert",
-  "face",
-  "fact",
-  "fall",
-  "false",
-  "family",
-  "farm",
-  "father",
-  "fear",
-  "feather",
-  "feeble",
-  "feeling",
-  "female",
-  "fertile",
-  "fiction",
-  "field",
-  "fight",
-  "finger",
-  "fire",
-  "first",
-  "fish",
-  "fixed",
-  "flag",
-  "flame",
-  "flat",
-  "flight",
-  "floor",
-  "flower",
-  "fold",
-  "food",
-  "foolish",
-  "foot",
-  "force",
-  "fork",
-  "form",
-  "forward",
-  "fowl",
-  "frame",
-  "free",
-  "frequent",
-  "friend",
-  "from",
-  "front",
-  "fruit",
-  "full",
-  "future",
-  "garden",
-  "general",
-  "girl",
-  "give",
-  "glass",
-  "glove",
-  "goat",
-  "gold",
-  "good",
-  "government",
-  "grain",
-  "grass",
-  "great",
-  "green",
-  "grey",
-  "grip",
-  "group",
-  "growth",
-  "guide",
-  "hair",
-  "hammer",
-  "hand",
-  "hanging",
-  "happy",
-  "harbour",
-  "hard",
-  "harmony",
-  "hate",
-  "have",
-  "head",
-  "healthy",
-  "hear",
-  "hearing",
-  "heart",
-  "heat",
-  "help",
-  "high",
-  "history",
-  "hole",
-  "hollow",
-  "hook",
-  "hope",
-  "horn",
-  "horse",
-  "hospital",
-  "hour",
-  "house",
-  "humour",
-  "idea",
-  "important",
-  "impulse",
-  "increase",
-  "industry",
-  "insect",
-  "instrument",
-  "insurance",
-  "interest",
-  "invention",
-  "iron",
-  "island",
-  "jelly",
-  "jewel",
-  "join",
-  "journey",
-  "judge",
-  "jump",
-  "keep",
-  "kettle",
-  "kick",
-  "kind",
-  "kiss",
-  "knee",
-  "knife",
-  "knot",
-  "knowledge",
-  "land",
-  "language",
-  "last",
-  "late",
-  "laugh",
-  "lead",
-  "leaf",
-  "learning",
-  "leather",
-  "left",
-  "letter",
-  "level",
-  "library",
-  "lift",
-  "light",
-  "like",
-  "limit",
-  "line",
-  "linen",
-  "liquid",
-  "list",
-  "little",
-  "living",
-  "lock",
-  "long",
-  "look",
-  "loose",
-  "loss",
-  "loud",
-  "love",
-  "machine",
-  "make",
-  "male",
-  "manager",
-  "mark",
-  "market",
-  "married",
-  "mass",
-  "match",
-  "material",
-  "meal",
-  "measure",
-  "meat",
-  "medical",
-  "meeting",
-  "memory",
-  "metal",
-  "middle",
-  "military",
-  "milk",
-  "mind",
-  "mine",
-  "minute",
-  "mist",
-  "mixed",
-  "money",
-  "monkey",
-  "month",
-  "moon",
-  "morning",
-  "mother",
-  "motion",
-  "mountain",
-  "mouth",
-  "move",
-  "much",
-  "muscle",
-  "music",
-  "nail",
-  "name",
-  "narrow",
-  "nation",
-  "natural",
-  "near",
-  "necessary",
-  "neck",
-  "need",
-  "needle",
-  "nerve",
-  "news",
-  "night",
-  "noise",
-  "normal",
-  "north",
-  "nose",
-  "note",
-  "number",
-  "observation",
-  "offer",
-  "office",
-  "only",
-  "open",
-  "operation",
-  "opinion",
-  "opposite",
-  "orange",
-  "order",
-  "organization",
-  "ornament",
-  "other",
-  "oven",
-  "over",
-  "owner",
-  "page",
-  "pain",
-  "paint",
-  "paper",
-  "parallel",
-  "parcel",
-  "part",
-  "past",
-  "paste",
-  "payment",
-  "peace",
-  "pencil",
-  "person",
-  "physical",
-  "picture",
-  "pipe",
-  "place",
-  "plane",
-  "plant",
-  "plate",
-  "play",
-  "please",
-  "pleasure",
-  "plough",
-  "pocket",
-  "point",
-  "poison",
-  "polish",
-  "political",
-  "poor",
-  "porter",
-  "position",
-  "possible",
-  "potato",
-  "powder",
-  "power",
-  "present",
-  "price",
-  "print",
-  "prison",
-  "private",
-  "probable",
-  "process",
-  "produce",
-  "profit",
-  "property",
-  "prose",
-  "protest",
-  "public",
-  "pull",
-  "pump",
-  "punishment",
-  "purpose",
-  "push",
-  "quality",
-  "question",
-  "quick",
-  "quiet",
-  "quite",
-  "rail",
-  "rain",
-  "range",
-  "rate",
-  "reaction",
-  "reading",
-  "ready",
-  "reason",
-  "receipt",
-  "record",
-  "regret",
-  "regular",
-  "relation",
-  "religion",
-  "representative",
-  "request",
-  "respect",
-  "responsible",
-  "rest",
-  "reward",
-  "rhythm",
-  "rice",
-  "right",
-  "ring",
-  "river",
-  "road",
-  "roll",
-  "roof",
-  "room",
-  "root",
-  "rough",
-  "round",
-  "rule",
-  "safe",
-  "sail",
-  "salt",
-  "same",
-  "sand",
-  "scale",
-  "school",
-  "science",
-  "scissors",
-  "screw",
-  "seat",
-  "second",
-  "secret",
-  "secretary",
-  "seed",
-  "seem",
-  "selection",
-  "self",
-  "send",
-  "sense",
-  "separate",
-  "serious",
-  "servant",
-  "shade",
-  "shake",
-  "shame",
-  "sharp",
-  "sheep",
-  "shelf",
-  "ship",
-  "shirt",
-  "shock",
-  "shoe",
-  "short",
-  "shut",
-  "side",
-  "sign",
-  "silk",
-  "silver",
-  "simple",
-  "sister",
-  "size",
-  "skin",
-  "skirt",
-  "sleep",
-  "slip",
-  "slope",
-  "slow",
-  "small",
-  "smash",
-  "smell",
-  "smile",
-  "smoke",
-  "smooth",
-  "snake",
-  "sneeze",
-  "snow",
-  "soap",
-  "society",
-  "sock",
-  "soft",
-  "solid",
-  "some",
-  "song",
-  "sort",
-  "sound",
-  "soup",
-  "south",
-  "space",
-  "spade",
-  "special",
-  "sponge",
-  "spoon",
-  "spring",
-  "square",
-  "stage",
-  "stamp",
-  "star",
-  "start",
-  "statement",
-  "station",
-  "steam",
-  "steel",
-  "stem",
-  "step",
-  "stick",
-  "sticky",
-  "stiff",
-  "still",
-  "stitch",
-  "stocking",
-  "stomach",
-  "stone",
-  "stop",
-  "store",
-  "story",
-  "straight",
-  "strange",
-  "street",
-  "stretch",
-  "strong",
-  "structure",
-  "substance",
-  "such",
-  "sudden",
-  "sugar",
-  "suggestion",
-  "summer",
-  "support",
-  "surprise",
-  "sweet",
-  "swim",
-  "system",
-  "table",
-  "tail",
-  "take",
-  "talk",
-  "tall",
-  "taste",
-  "teaching",
-  "tendency",
-  "test",
-  "than",
-  "that",
-  "then",
-  "theory",
-  "there",
-  "thick",
-  "thin",
-  "thing",
-  "this",
-  "thought",
-  "thread",
-  "throat",
-  "through",
-  "through",
-  "thumb",
-  "thunder",
-  "ticket",
-  "tight",
-  "till",
-  "time",
-  "tired",
-  "together",
-  "tomorrow",
-  "tongue",
-  "tooth",
-  "touch",
-  "town",
-  "trade",
-  "train",
-  "transport",
-  "tray",
-  "tree",
-  "trick",
-  "trouble",
-  "trousers",
-  "true",
-  "turn",
-  "twist",
-  "umbrella",
-  "under",
-  "unit",
-  "value",
-  "verse",
-  "very",
-  "vessel",
-  "view",
-  "violent",
-  "voice",
-  "waiting",
-  "walk",
-  "wall",
-  "warm",
-  "wash",
-  "waste",
-  "watch",
-  "water",
-  "wave",
-  "weather",
-  "week",
-  "weight",
-  "well",
-  "west",
-  "wheel",
-  "when",
-  "where",
-  "while",
-  "whip",
-  "whistle",
-  "white",
-  "wide",
-  "will",
-  "wind",
-  "window",
-  "wine",
-  "wing",
-  "winter",
-  "wire",
-  "wise",
-  "with",
-  "woman",
-  "wood",
-  "wool",
-  "word",
-  "work",
-  "worm",
-  "wound",
-  "writing",
-  "wrong",
-  "year",
-  "yellow",
-  "yesterday",
-  "young",
-];
+const dictionary = require("./dictionary");
+
+let gameObject = {
+  gameBoard: {},
+  guessesLeft: 7,
+  guessedLetters: [],
+  guessCounter: 0,
+};
+
 // get random word
-let secretWord = "";
 const getWord = () => {
   secretWord = dictionary[Math.floor(Math.random() * dictionary.length)];
   // console.log("\nThe secret word is: " + secretWord);
   return secretWord;
 };
-getWord();
 
-let gameBoard = {};
-let guessCounter = 0;
-let guessesLeft = 7;
-let guessedLetters = [];
-let lettersRemaining = secretWord.length;
+// Start game function - will reset the game if you want to play again
+const startGame = () => {
+  console.clear();
+  gameObject["guessesLeft"] = 7;
+  gameObject["guessedLetters"] = [];
+  gameObject["gameBoard"] = {};
+  gameObject["guessCounter"] = 0;
 
+  getWord();
+  lettersRemaining = secretWord.length;
+
+  welcomeMessage();
+  displayBoard();
+
+  while (gameObject["guessesLeft"] > 0 && lettersRemaining > 0) {
+    const guess = getValidLetterGuess();
+    if (!hasLetterBeenGuessedAlready(guess)) {
+      updateGameBoard(guess);
+      deincrementGuess(guess);
+      guessedLettersArray(guess);
+      winGame(lettersRemaining);
+      gameOver(gameObject["guessesLeft"]);
+    }
+  }
+};
+
+// Displays the opening welcome message
 const welcomeMessage = () => {
-console.log("Welcome to Snowman! You have 7 guesses - Enjoy!\n");
-}
-welcomeMessage();
+  console.log("Welcome to Snowman! You have 7 guesses - Enjoy!\n");
+};
 
 // The loop creating the answer array - will update with correct guesses
 const displayBoard = () => {
-for (let i = 0; i < secretWord.length; i++) {
-  gameBoard[i] = "_";
-}
-console.log(Object.values(gameBoard).join(" "));
-}
-displayBoard();
+  for (let i = 0; i < secretWord.length; i++) {
+    gameObject.gameBoard[i] = "_";
+  }
+  console.log(Object.values(gameObject.gameBoard).join(" "));
+};
 
 // deincrement guess
 const deincrementGuess = (letter) => {
-  if (guessedLetters.includes(letter)) {
-    console.log(
-      "YOU ALREADY GUESSED THAT LETTER - PLEASE TRY A DIFFERENT GUESS!"
-    );
+  if (secretWord.includes(letter)) {
+    console.log("\nGood guess!");
   } else {
-    guessesLeft--;
-    console.log("\nYou have " + guessesLeft + " guesses remaining\n");
+    gameObject["guessesLeft"]--;
   }
+  console.log(
+    "\nYou have " + gameObject["guessesLeft"] + " guesses remaining\n"
+  );
 };
 
 // adds guessed letters to array and displays it
 const guessedLettersArray = (letter) => {
-  guessedLetters.push(letter);
-  console.log("\nGuessed letters: " + guessedLetters);
+  gameObject["guessedLetters"].push(letter);
+  console.log("\nGuessed letters: " + gameObject["guessedLetters"]);
 };
 
 // updates gameboard if correct letter chosen
@@ -796,56 +77,34 @@ const updateGameBoard = (letter) => {
   console.clear();
   for (k = 0; k < secretWord.length; k++) {
     if (secretWord[k] === letter) {
-      gameBoard[k] = letter;
+      gameObject.gameBoard[k] = letter;
       lettersRemaining--;
-      guessesLeft++;
     }
   }
-  console.log(Object.values(gameBoard).join(" "));
-  // console.log("\nLetters Remaining = " + lettersRemaining);
+  console.log(Object.values(gameObject.gameBoard).join(" "));
 };
 
 // displays win game message when players guesses word correctly
 const winGame = (lettersRemaining) => {
   if (lettersRemaining === 0) {
     console.log("\nCongratulations YOU HAVE WON!!");
-    console.log("\nIt took you " + guessCounter + " guesses!");
+    console.log("\nIt took you " + gameObject["guessCounter"] + " guesses!");
     console.log("\nThe word was: " + secretWord);
     if (readLineSync.keyInYN("\nWould you like to try again?")) {
-      resetGame();
-      getValidLetterGuess();
+      startGame();
     } else {
       process.exit();
     }
   }
 };
 
-// resets the game if players wants to play again
-const resetGame = () => {
-  getWord();
-  guessesLeft = 7;
-  guessedLetters = [];
-  gameBoard = {};
-  guessCounter = 0;
-  lettersRemaining = secretWord.length;
-  console.clear();
-  welcomeMessage();
-  displayBoard();
-  getValidLetterGuess();
-
-  // let secretWord = dictionary[Math.floor(Math.random() * dictionary.length)];
-  // // console.log("\nThe secret word is: " + secretWord);
-  // lettersRemaining = secretWord.length;
-};
-
 // displays game over message when player runs out of guesses
-const gameOver = (guessesLeft) => {
-  if (guessesLeft === 0 && lettersRemaining !== 0) {
+const gameOver = () => {
+  if (gameObject["guessesLeft"] === 0 && lettersRemaining !== 0) {
     console.log("\nSorry you ran out of guesses! Better luck next time!");
     console.log("\nThe word was: " + secretWord);
     if (readLineSync.keyInYN("\nWould you like to try again?")) {
-      resetGame();
-      // getValidLetterGuess();
+      startGame();
     } else {
       process.exit();
     }
@@ -854,9 +113,8 @@ const gameOver = (guessesLeft) => {
 
 // has letter already been guessed function
 const hasLetterBeenGuessedAlready = (letter) => {
-  for (let p = 0; p < guessedLetters.length; p++) {
-    if (guessedLetters[p] === letter) {
-      // guessesLeft++;
+  for (let p = 0; p < gameObject["guessedLetters"].length; p++) {
+    if (gameObject.guessedLetters[p] === letter) {
       console.log(
         "\n YOU ALREADY GUESSED THAT LETTER - PLEASE TRY A DIFFERENT GUESS"
       );
@@ -866,34 +124,23 @@ const hasLetterBeenGuessedAlready = (letter) => {
   return false;
 };
 
-// the game loop
-while (guessesLeft > 0) {
-  function getValidLetterGuess() {
-    function guessIsValid(letter) {
-      return (
-        letter.length === 1 && letter.toUpperCase() != letter.toLowerCase()
-      );
-    }
-    let letter = "";
-    while (!letter) {
-      let input = readline.question("\nPlease enter your guess: ");
-      if (guessIsValid(input)) {
-        letter = input.toLowerCase();
-
-        guessCounter++;
-      } else {
-        console.log("\nPlease enter a valid letter");
-      }
-    }
-    const letterGuess = hasLetterBeenGuessedAlready(letter);
-    if (letterGuess === false) {
-      updateGameBoard(letter);
-      deincrementGuess(guessesLeft, letter);
-      guessedLettersArray(letter);
-      winGame(lettersRemaining);
-      gameOver(guessesLeft);
-    }
-    return letter.toLowerCase();
+// Get a valid letter guess
+function getValidLetterGuess() {
+  function guessIsValid(letter) {
+    return letter.length === 1 && letter.toUpperCase() != letter.toLowerCase();
   }
-  getValidLetterGuess();
+  let letter = "";
+  while (!letter) {
+    let input = readLineSync.question("\nPlease enter your guess: ");
+    if (guessIsValid(input)) {
+      letter = input.toLowerCase();
+
+      gameObject["guessCounter"]++;
+    } else {
+      console.log("\nPlease enter a valid letter");
+    }
+  }
+
+  return letter.toLowerCase();
 }
+startGame();
